Add core dispersive material classes
Analytical analysis of the optical properties of materials can prove
difficult due to the wide variety of materials of interest and their
non-trivial dispersion relations. As such, numerical analysis is often
preferred. To this extent, the Material class has been developed to
encapsulate the optical properties of a material in a manner convenient
for numerical calculation.

The differences in properties between optical materials can chiefly be
described by their different dispersion relations. An excellent database
of these dispersion relations is the RefractiveIndex.info database,
curated by Mikhail Polyanskiy. The RIIDMaterial class has been developed
to provide an easy interface to this database and it's wide assortment
of optical materials.

- Added Material class
- Added RIIDMaterial class
- Added basic documentation